requirements.txt
bytetracker==0.3.2
coloredlogs==15.0.1
filelock==3.25.0
flatbuffers==25.12.19
fsspec==2026.2.0
humanfriendly==10.0
Jinja2==3.1.6
lapx==0.9.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.25.2
onnxruntime-gpu==1.19.0
opencv-python==4.13.0.92
packaging==26.0
protobuf==7.34.0
pyreadline3==3.5.4
scipy==1.9.3
sympy==1.13.1
torch==2.5.1+cu121
typing_extensions==4.15.0
